Compile every class from source; restore the last 8 damaged classes
All app classes are now compiled from source and verified against the
3.4.2 bytecode (0 damaged, 0 unresolved references). bytecode_classes.txt
is empty and app/libs/lumiya-original-classes.jar is removed.

Repairs, each checked against the original smali:
- GvrSurfaceView.GLThread.guardedRun: jadx, CFR and Vineflower all failed on
  it, and the stub threw UnsupportedOperationException, so the VR render
  thread died on start. Written by hand: AOSP GLSurfaceView render loop plus
  Google VR's swap-mode handling (EGL_RENDER_BUFFER, front-buffer
  auto-refresh; swap skipped in single-buffer mode).
- TextureCompressedCache.TextureFetchRequest.run and the
  SLInventoryHTTPFetchRequest fetch runnable: back to 3.4.2 behaviour
  (2 / 3 attempts, original logging, printStackTrace paths).
- SLInventoryHTTPFetchRequest.DatabaseCommitThread.run: the earlier rewrite
  used one transaction for the whole fetch and dropped the per-batch commit
  and yieldIfContendedSafely() every 16 entries. Restored.
- UserFunctionsFragment: four dialogs back on android.app.AlertDialog.
- SyncManager.syncMoreMessages: explicit StringBuilder, as in 3.4.2.
- SLSearch: Room mirror via SearchGridResultRepositoryAdapter (6f972bd)
  kept and recorded in accepted.txt.

Verifier (tools/verify/dexdiff.py):
- field references are bound to the declaring class, so
  SLObjectAvatarInfo->treeNode matches SLObjectInfo->treeNode;
- primitive-array clone() calls are no longer dropped from fingerprints;
- methods whose losses were excused by a removed SDK_INT check are now
  listed (SDKFOLD) for review instead of being passed silently. Checking
  these by hand found the DatabaseCommitThread loss above.

Readability: tools/recover/names/FixSwitchMaps.java replaces the jadx
switch-map helpers (switch (m199x...()[k.ordinal()]) { case 1: ... }) with
plain enum switches in 35 files.

Docs: verified_recovery.md, superseded_hand_edits.txt, BUILD_STATUS.md.

+++ app/src/main/java/com/lumiyaviewer/lumiya/ui/common/ImageAssetView.java
@@ -145,7 +145,10 @@ public class ImageAssetView extends View {
         }
         int width2 = this.imageBitmap.getWidth();
         int height2 = this.imageBitmap.getHeight();
-        float max = Math.max(width2 / width, height2 / height);
+        // Fit the bitmap inside the view: scale by the larger of the two ratios.
+        // Must be float division (the decompiled source used int division, which
+        // made the scale 0 for any image smaller than the view).
+        float max = Math.max((float) width2 / (float) width, (float) height2 / (float) height);
         int round = Math.round(width2 / max);
         int scaledHeight = Math.round(height2 / max);
         int i = (width / 2) - (round / 2);
